requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
langsmith
langgraph-checkpoint
pydantic
python-dotenv
trustcall
tavily-python
beautifulsoup4
wikipedia
arxiv 
pymupdf
chromadb
streamlit
langchain-chroma